I think I got the locks wrong, fix deadlock.
I let the breakEverything script run overnight and found it deadlocked after a
few hours. I watched it run for a while, and think WorldRenderLayer::Render and
Context::Update are the best places to inject locks.

So far it seems to not crash or deadlock, but it's possible this is still
wrong. I still need to add them to the other IntersectManager routines, but need
to write a testcase to make sure they work.